Avoid string copying trick, refactor hot loop
@alaviss offered a trick which avoids superfluous copying of a string
on slicing the buffer and table insertion. This alone boosted the speed
by ~20%. Standard library tables need to take advantage of Nim's
move semantics.

The main loop was rewritten to use a basic FSM and a few buggy hectic
lines were put in order so everything is a bit simpler and shorter.
Unfortunately, only negligible speed gains (<5%) on this front.